+++ stratapy/__init__.py
@@ -3,7 +3,7 @@ stratapy: A Python package for stratigraphic plotting and analysis.
 
 Documentation: https://stratapy.readthedocs.io/en/latest/
 
-Publication: https://doi.org/10.5281/zenodo.19320013
+Publication: https://doi.org/10.1038/s41598-026-58501-2
 
 Authors: Jack Lee Smith, Christina Antoniou, Ruaridh Alexander
 """