+++ stratapy/__init__.py
@@ -9,7 +9,7 @@ Authors: Jack Lee Smith, Christina Antoniou, Ruaridh Alexander
 """
 
 __all__ = ['update_minerals', 'update_lithologies', 'update_features', 'update_contacts', 'load', 'chronostratigraphy', 'multi_fig', 'correlated_logs', 'standalone_legend', 'list_examples']
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 __author__ = 'Jack Lee Smith, Christina Antoniou, Ruaridh Alexander'
 __license__ = 'BSD 3-Clause License'
 __copyright__ = 'Copyright (c) 2026 Jack Lee Smith'